Small fixes: NSInteger cast to long, empty initWithTelemetryString method removed from protocol and objects

+++ IdentityCore/src/telemetry/request_telemetry/MSIDLastRequestTelemetrySerializedItem.m
@@ -69,7 +69,7 @@
         int lastIndex = (int)self.errorsInfo.count - 1;
         
         // Set first post in fencepost structure -- last item in string doesn't have comma at the end
-        NSString *currentFailedRequest = [NSString stringWithFormat:@"%ld,%@", self.errorsInfo[lastIndex].apiId, self.errorsInfo[lastIndex].correlationId ?: @""];
+        NSString *currentFailedRequest = [NSString stringWithFormat:@"%ld,%@", (long)self.errorsInfo[lastIndex].apiId, self.errorsInfo[lastIndex].correlationId ?: @""];
         NSString *currentErrorMessage = [NSString stringWithFormat:@"%@", self.errorsInfo[lastIndex].error ?: @""];
         
         // Only add error info into string if the resulting string smaller than 4KB
